requirements.txt
langgraph
langchain
chromadb
networkx
pydantic
numpy
pandas
pytest
pyvis
colorama
python-json-logger
openai
python-dotenv
rouge-score
scikit-learn
sentence-transformers
bert-score
nltk
jinja2
httpx
tqdm
nltk

